+++ docs/additional_information/changelog.rst
@@ -24,6 +24,7 @@ Added
     or labels
 - Added the property ``__version__`` to ``dtaianomaly``, which can be accessed from code.
 - Included the used version of ``dtaianomaly`` when logging errors.
+- Added the ``plot_time_series_anomalies`` function to visualize time series data with real and predicted anomalies, highlighting TP, FP, and FN cases.
 - Implemented ``PrincipalComponentAnalysis``, ``KernelPrincipalComponentAnalysis`` and
   ``RobustPrincipalComponentAnalysis`` anomaly detectors.
 - Implemented ``HistogramBasedOutlierScore`` anomaly detector.